requirements.txt
flask
psycopg2-binary
sqlalchemy
pyyaml
flask-sqlalchemy
pyopenssl
flask_httpauth
pyjwt
werkzeug